refactor: Restructure models into separate modules for better maintainability
- Split monolithic models.py into modular files: `user.py`, `account.py`, `transaction.py`, `category.py`, `goal.py`, and `financial_plan.py`.
- Added a central `models/__init__.py` for streamlined imports.
- Preserved all functionality and relationships across modules.
