+++ app/models/__init__.py
@@ -1,0 +1,6 @@
+from .user import User
+from .account import Account
+from .transaction import Transaction
+from .category import Category
+from .goal import Goal
+from .financial_plan import FinancialPlan
