+++ app/models/goal.py
@@ -1,0 +1,20 @@
+from decimal import Decimal
+from .. import db
+
+class Goal(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    name = db.Column(db.String(120), nullable=False)
+    target_amount = db.Column(db.Numeric(precision=10, scale=2), nullable=False)
+    user_id = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
+    account_id = db.Column(db.Integer, db.ForeignKey('account.id'), nullable=False)
+    account = db.relationship('Account', backref='goal', uselist=False)
+
+    @property
+    def current_amount(self):
+        return self.account.balance if self.account else Decimal('0.00')
+
+    @property
+    def progress_percentage(self):
+        if self.target_amount > 0:
+            return (self.current_amount / self.target_amount) * 100
+        return 0
